fix(dev-test): fix error in pytest setting path in pyroject.toml, remove from .vscode settings. + Otimize settings
was incorretly set as tool.pytest and not tool.pytest.ini_options

After making it work in there, should not be keept in .vscode settings as it's not vscode specific.

Optimized test collection parameters to not recurse some typically large dirs. Might not make a difference as testpaths is set, but dosen't hurt and seems to speed upp collection.

+ Log test call durations
+ When running begin by running last modified scripts