refactor(environment): remove sqlite dependency from environment.dev.yml
Removed the sqlite dependency from the environment.dev.yml file as it was unnecessary for the current project requirements. This helps to streamline the dependencies and reduce potential conflicts in the environment setup.

Sqlite was intoduced on a previous iteratation, but is currently not used. If introduced it should be explicitly added.